Fixed the in-place layers so that they don't interfere with the operation of skip layers and add_prev style layers.  In particular, now in-place layers only overwrite the gradient information in their child layer if they are operating in in-place mode.  Otherwise, they add their gradients to their child layers.
It should also be noted that it's safe for in-place layers to overwrite
gradients when in in-place mode since their child layers are inaccessible when
in-place layers operate in in-place mode.  This prevents any other layers from
trying to add to the child layer, thereby avoiding the potability of layer
interference.  So the bug this change fixes is that, when not in in-place mode
the child layers are still accessible but in-place layers were *still*
overwriting child gradients.