Add new proxies for serializing to std::ostream and deserializing from std::istream.
I wanted to do as follows:

    net_type net, net2;
    // ...
    std::ostringstream out;
    serialize(net, out);
    const std::string serialized = out.str();
    // ...
    std::istringstream in(serialized);
    deserialize(net2, in);

(In other words, I wanted to serialize DNNs, but in memory and not via files.)

I couldn't figure out a way to do this without adding this code somewhere: otherwise, MSVC++ would just complain that it can't match net and net2 (didn't try other compilers). First added the new code on the call site, but as it apparently has to be in the dlib namespace, thought it'd be better to have it here.

Perhaps I was doing it all wrong to begin with, but at least this addition did let me do what I wanted to do, so thought I'd share it.