Added SARSA + Qlearning RL algorithms
I've implemented the RL algorithms QLearning and SARSA. Both of them are
quite similar to the lspi code and I might try to merge all their common
parts in a template superclass.

Added and passed tests for both of them based in the same case that
lspi.
At the moment they don't manage the case when the last feature has to be
1 (it is declared like so in the contract).