Automatically test for NEON instructions when building python extensions and enable them if available.
Forgot about --yes USE_NEON_INSTRUCTIONS when removing --yes a moment
ago.  It's on by default now so now --yes really is unneeded.  The dlib
policy on this kind of thing going forward is to always have a cmake
script that tests if a feature like this is available and to enable it
automatically when building python extensions.