add iou_anchor_threshold
This allows changing the behavior of the network while training.
Setting it to 1 (disable) will have the standard YOLO behavior:
match only the best anchor with a truth.  However, setting it to
a lower value will make any anchor with an IoU > iou_anchor_threshold
with a ground truth be updated as well.
See this thread for more details:
https://github.com/AlexeyAB/darknet/issues/7087