Fix dnn_trainer trying to decrease the learning rate
If the learnin_rate_shrink == 1 (we have a custom learning rate schedule),
the trainer will tell us that it will decrease the learning rate, when the
loss is increasing, but in reality it won't.  I am not sure how to properly
address this case, but at least  it shouldn't lie about it.