We have some excessive and duplicative tests in the travis-ci setup.
This is causing us to run out of travis-ci credits, making tests not run
at all.  I deleted the duplicative tests and then disabled two additonal
ones by commenting them out that would be nice to run but I think are
not essential.  In particular, the OSX one eats up a ton of credits.  So
I disabled that.  Maybe we can turn it back on later if we end up well
under the credit budget (or switch to github actions which appears to
have higher limits)