Problem: `Error while calling cudnnGetConvolutionForwardWorkspaceSize( context(), descriptor(data), (const cudnnFilterDescriptor_t)filter_handle, (const cudnnConvolutionDescriptor_t)conv_handle, descriptor(dest_desc), (cudnnConvolutionFwdAlgo_t)forward_algo, &forward_workspace_size_in_bytes) in file C:\a\2\s\3rdparty\dlib\dlib\cuda\cudnn_dlibapi.cpp:1029. code: 9, reason: CUDNN_STATUS_NOT_SUPPORTED`
Solution: Add remaining dimensions of `filters`, and all dimensions of `data` to cache key

Not sure, but it looks like otherwise the chosen algorithm may be unsupported for some input combinations (e.g., larger input?).

Also not sure if all those key components are really needed. But reproducing the problem is not very straightforward, so figuring that out would take some time. So personally would be fine to just err on the side of caution. The present solution at least seems to fix the problem.

An alternative, I guess, could have been to try out `cudnnGetConvolutionForwardWorkspaceSize`, and in the event of failure, just clean the related cache item and try again.